[Fix] Force the sigmoid calculation in multi-label head to be full-precision. (#1251)
* [CI] Update CI of the master branch. (#1252)

* [Bot] Update assignee schedule. (#1262)

* [Fix] Force the sigmoid calculation in multi-label head to be
full-precision.

+++ mmcls/models/heads/multi_label_head.py
@@ -82,7 +82,8 @@ class MultiLabelClsHead(BaseHead):
             x = x[-1]
 
         if sigmoid:
-            pred = torch.sigmoid(x) if x is not None else None
+            # Convert to full precision because sigmoid is sensitive.
+            pred = torch.sigmoid(x.float()) if x is not None else None
         else:
             pred = x
